Refactor the test
